Implement jumps
As a part of implementing jumps, I've also added the foundation for
weighted cladogenesis types. Currently, the weights are working, but
incorrectly. I still need to implement a method of counting the
different split types for a given distribution.

Additionally, we have changed to the rejection method to ensure that the
results are correct. I will probably keep this implemenation forever for
the purposes of testing.