Refactor the regression tests, and fix some bugs
This is kinda a messy commit. There are also a bunch of refactor changes
in here.

Regarding the bugs with regression, it was the case that we were
classifying too many of the random splits as jumps, so we had to fix
that. In some sense, this is "fixing the tests", but having two separate
methods that agree is certainly good.

Additionally, there is a weird duplicity requirement for singleton. To
make the probability work out for the region `1000` with jumps, we have
to double count the copy event.

I still have to get duplicity working officially. I think that this is
gonna be complicated.